pyproject.toml
[tool.poetry]
name = "polar-hr"
version = "0.1.0"
description = ""
authors = ["Nina <280141+ninas@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "polar_hr"}]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^2.6.0"
google-cloud-secret-manager = "^2.12.6"
peewee = "^3.15.4"
overrides = "^7.3.1"
isodate = "^0.6.1"
psycopg2-binary = "^2.9.5"
google-api-python-client = "^2.66.0"
click = "^8.1.3"
requests = "^2.28.1"
requests-oauthlib = "^1.3.1"
google-cloud-runtimeconfig = "^0.33.2"
google-cloud-logging = "^3.3.1"
structlog = "^22.3.0"
gcloud-aio-storage = "^7.0.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
functions-framework = "^3.2.1"
rich = "^12.6.0"
isort = "^5.11.1"

[tool.isort]
profile = "black"
skip = [".gitignore", ".dockerignore"]
extend_skip = [".cache", "node_modules"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

